+++ tests/testthat/test_operations_setpower.R
@@ -1,11 +1,8 @@
-library(testthat)
-
-context("setpower")
-
 test_that("Set", {
   expect_equal(setpower(Set$new(1, 2), 2, simplify = FALSE)$strprint(), "{1, 2}^2")
   expect_true(setpower(Set$new(1, 2), 2, simplify = TRUE)$equals(Set$new(Tuple$new(1, 1), Tuple$new(2, 1), Tuple$new(1, 2), Tuple$new(2, 2))))
   expect_equal(Set$new(1, 2)^1, Set$new(1, 2))
+  expect_equal(Set$new(1, 2)^0, Set$new())
 })
 
 test_that("conditionalset", {
@@ -31,4 +28,13 @@ test_that("setwrapper", {
     "([1,2] U [3,4])^2"
   )
   useUnicode(TRUE)
+})
+
+test_that("variable", {
+  x <- Interval$new(0, 1)^"n"
+  expect_true(x$contains(Tuple$new(0)))
+  expect_true(x$contains(Tuple$new(0, 1)))
+  expect_true(x$contains(Tuple$new(0, 1, 0, 0, 1, 1, 0)))
+  expect_false(x$contains(Tuple$new(0, 2)))
+  expect_error(x$contains(1))
 })