+++ NEWS.md
@@ -3,6 +3,7 @@
 * Bugfix in set operation cleaner
 * Bugfix causing `Interval$contains` to return `TRUE` for tuples
 * Bugfix in union sets incorrectly unwrapping products
+* Added variable length `ExponentSet`s
 
 # set6 0.1.5